Start to add integration of potent actions
This patch begins to add "potent" operations to the API. These are operations like "object_put," etc. As this is the best implemented IPFS implementation in Elixir, I believe it's worth building the rest of the API beyond what's currently available (that is, code to read from IPFS). We begin with some simple functions for minipulating objects.